Issue #142: Add THC for useful methods in testing
