+++ app/src/androidTest/java/com/macbitsgoa/ard/activities/AuthActivityTest.java
@@ -78,25 +78,4 @@ public class AuthActivityTest {
                     .check(matches(isDisplayed()));
         }
     }
-
-    public class ToastMatcher extends TypeSafeMatcher<Root> {
-
-        @Override
-        public void describeTo(Description description) {
-            description.appendText("is toast");
-        }
-
-        @Override
-        public boolean matchesSafely(Root root) {
-            int type = root.getWindowLayoutParams().get().type;
-            if ((type == WindowManager.LayoutParams.TYPE_TOAST)) {
-                IBinder windowToken = root.getDecorView().getWindowToken();
-                IBinder appToken = root.getDecorView().getApplicationWindowToken();
-                if (windowToken == appToken) {
-                    //means this window isn't contained by any other windows.
-                }
-            }
-            return false;
-        }
-    }
 }
