+++ app/src/androidTest/java/com/macbitsgoa/ard/activities/MainActivityTest.java
@@ -40,6 +40,7 @@ import static android.support.test.espresso.matcher.ViewMatchers.withEffectiveVi
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withParent;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
+import static com.macbitsgoa.ard.THC.childAtPosition;
 import static org.hamcrest.Matchers.allOf;
 
 /**
@@ -114,26 +115,6 @@ public class MainActivityTest {
         onView(withId(R.id.fragment_chat_layout)).check(matches(isDisplayed()));
         onView(withId(R.id.fragment_home_layout)).check(doesNotExist());
         onView(withId(R.id.fragment_faq_layout)).check(doesNotExist());
-    }
-
-    private static Matcher<View> childAtPosition(
-            final Matcher<View> parentMatcher, final int position) {
-
-        return new TypeSafeMatcher<View>() {
-            @Override
-            public void describeTo(Description description) {
-                description.appendText("Child at position " + position + " in parent ");
-                parentMatcher.describeTo(description);
-            }
-
-            @Override
-            public boolean matchesSafely(View view) {
-                ViewParent parent = view.getParent();
-                return parent instanceof ViewGroup && parentMatcher.matches(parent)
-                        && view.equals(((ViewGroup) parent).getChildAt(position));
-            }
-        };
-
     }
 
     @Test
