+++ app/src/androidTest/java/com/macbitsgoa/ard/fragments/HomeAnnFragmentTest.java
@@ -37,6 +37,8 @@ import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.withEffectiveVisibility;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
+import static com.macbitsgoa.ard.THC.getString;
+import static com.macbitsgoa.ard.THC.hasTextColor;
 import static org.hamcrest.CoreMatchers.allOf;
 
 /**
@@ -63,7 +65,7 @@ public class HomeAnnFragmentTest {
     public void testLimitCrossed() throws Exception {
         uiDevice.setOrientationNatural();
         //Get input from text file
-        final String text = getString("lipsum-big.txt");
+        final String text = getString(getClass(), "lipsum-big.txt");
 
         onView(withId(R.id.editText_home_ann_fragment_message)).perform(typeText(text));
         pressBack();
@@ -91,7 +93,7 @@ public class HomeAnnFragmentTest {
     @Test
     public void testLimitNotCrossed() throws Exception {
         //Get input from text file
-        final String text = getString("lipsum-small.txt");
+        final String text = getString(getClass(), "lipsum-small.txt");
 
         onView(withId(R.id.editText_home_ann_fragment_message)).perform(typeText(text));
 
@@ -117,7 +119,7 @@ public class HomeAnnFragmentTest {
         onView(withId(R.id.fab_activity_post)).perform(click());
         onView(allOf(withId(android.support.design.R.id.snackbar_text), withText("Please enter message!")))
                 .check(matches(isDisplayed()));
-        text = getString("lipsum-small.txt");
+        text = getString(getClass(), "lipsum-small.txt");
 
         onView(withId(R.id.editText_home_ann_fragment_message)).perform(typeText(text));
         pressBack();
@@ -125,45 +127,5 @@ public class HomeAnnFragmentTest {
         onView(withId(R.id.fab_activity_post)).perform(click());
     }
 
-    /**
-     * Returns a string object from an input stream.
-     * The resource file should be placed in the same package name as the java file
-     * expect it should be inside the <b>resources</b> package and not the java package.
-     *
-     * @param file File name to be used.
-     * @return converted string.
-     * @throws Exception thrown exception.
-     */
-    public String getString(final String file) throws Exception {
-        InputStream inputStream = getClass().getResourceAsStream(file);
-        final ByteArrayOutputStream result = new ByteArrayOutputStream();
-        byte[] buffer = new byte[1024];
-        int length;
-        while ((length = inputStream.read(buffer)) != -1) {
-            result.write(buffer, 0, length);
-        }
-        return result.toString("UTF-8");
-    }
 
-
-    /**
-     * Matcher for EditText/TextView color checker.
-     *
-     * @param color color to match.
-     * @return Matcher.
-     */
-    public static Matcher<View> hasTextColor(final int color) {
-        Checks.checkNotNull(color);
-        return new BoundedMatcher<View, TextView>(TextView.class) {
-            @Override
-            public void describeTo(final Description description) {
-                description.appendText("with text color: ");
-            }
-
-            @Override
-            public boolean matchesSafely(final TextView warning) {
-                return color == warning.getCurrentTextColor();
-            }
-        };
-    }
 }
